scraper-worker/requirements.txt
kafka-python==2.0.2
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.9  